refactor(component): создан компонент Player, а старый переименован в PlayerPage
