+++ src/components/pages/player-page/player-page.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+import MainLayout from "../../layouts/main-layout/main-layout";
+import Player from "../../sections/player/player";
+
+const PlayerPage = () => {
+
+  return (
+    <MainLayout>
+      <Player/>
+    </MainLayout>
+  );
+};
+
+export default PlayerPage;
